Added parameter to select the post-processing operation to apply to all bitmaps generated from a wxBitmapBundle containing SVG data.
When using bitmaps as icons in buttons or toolbars tipically an enabled and a disabled version of the same image are
required. The disabled version can be generated from the enabled one by using the ConvertToDisabled method, avoiding
the need to include two images per button in the executable.
When working with bitmap bundles this is more troublesome as the enabled bitmap needs to be created from bitmap data
first, the disabled bitmap has to be generated with the ConvertToDisabled method and then both bitmaps need to be
stored in separate bundles.
When working with SVGs even this is no longer possible as there's no method to modify the SVG data. And since the
buttons get the bitmap from the bundle for themselves, two versions of the same SVG need to be stored on the
executable. The new parameter allows the use of the factory methods like this:

	wxBitmapBundle active = wxBitmapBundle::FromSVG(data, wxSize(32, 32));
	All bitmaps generated from the SVG without post-processing

	wxBitmapBundle disabled = wxBitmapBundle::FromSVG(data, wxSize(32, 32), wxBitmapBundle::FromSVGDim);
	All bitmaps generated from this bundle will have their alpha channel at 33% of the original

	wxBitmapBundle active_dark = wxBitmapBundle::FromSVG(data, wxSize(32, 32), wxBitmapBundle::FromSVGInvertGreys);
	All bitmaps generated from this bundle will have their grey pixels inverted

	wxBitmapBundle disabled_dark = wxBitmapBundle::FromSVG(data, wxSize(32, 32), wxBitmapBundle::FromSVGDim |
                                                                                     wxBitmapBundle::FromSVGInvertGreys);
	All bitmaps generated from this bundle will have their alpha channel at 33% of the original
	and their grey pixels inverted

So now we can get four versions of the same SVG, saving up a lot of memory.