Fix grid window scrollbars when freezing part of the grid
The space taken by the frozen part needs to be subtracted from the
available size, not added to it, as was mistakenly done before.

This fixes the problem with scrollbars not appearing correctly in the
grid even when they were clearly needed because its contents didn't fit
into the window.

Closes #22602.

(cherry picked from commit c88d18278b182cbd8046d0231fe4db313d772a99)