Correct default OnCloseDocument() implementation documentation
Don't say that the version of this function in wxDocument checks whether
the document is modified and asks the user if it should be saved because
this is not the case: this is done in a separate OnSaveModified()
function called earlier.

Also document that this function should always return true.