Split the wxAuiToolBarArt::UpdateColoursFromSystem method into two.
This facilitates the use of different bitmaps or bundles as drop down and overflow icons,
without needing to replicate code. To specify the bitmaps to use a sub-class of
wxAuiGenericToolBarArt should be created, with a constructor that sets the intended width
of the bitmaps by setting the m_overflowSize and m_dropdownSize variables and by overriding
the UpdateBitmapColoursFromSystem method and setting up the bitmaps into the
m_buttonDropDownBmp, m_disabledButtonDropDownBmp, m_overflowBmp and m_disabledOverflowBmp
variables. An instance of that class is then passed into the wxAuiToolBar as the art provider.