BUGFIX: Moved the UpdateColoursFromSystem call out of the wxAuiGenericToolBarArt constructor
The UpdateColoursFromSystem method is declared virtual but being called from the constructor
prevents virtual dispatching the call into a sub-class overriden method, forcing the user
code to call it if a sub-class is to be used as art provider. Now the call is done when
the sub-class is fed into the wxAuiToolBar as an art provider.