Change on bc8c178e4f915e3b64e08d3c34559020bec16570; new event along with EVT_MENU on button click
The patch mentioned above required the use of the GetExtraLong() method of the wxCommandEvent
issued on a mouse click on the tool to find out if this click was done after using the dropdown
button to open an auxiliary window.
This change adds a new set of methods to the wxAuiToolBarEvent to report that information, named
IsButtonClickedAfterDropDown  and SetButtonClickedAfterDropDown, and issues an wxAuiToolBarEvent
right after the wxCommandEvent. Therefore now to detect button presses the used code can also
bind to wxAuiToolBarEvents with IDs of wxEVT_AUITOOLBAR_TOOL_BUTTON, if it wants to detect the
behaviour.