Fix using windows with sizers as wxStaticBoxSizer labels in XRC
The contents of <windowlabel> can be any window, including a panel with
its own sizer, but such panels were not created correctly due to using
the wrong value of wxSizerXmlHandler::m_parentSizer, which prevented the
sizer from being associated with the panel.

Fix this by using the same hack as elsewhere and temporarily resetting
the value of this variable before calling CreateResFromNode(). Also do
the same for m_isInside, even though it doesn't seem necessary here,
because it should always be false anyhow when <windowlabel> is parsed,
just for consistency. But don't bother doing it for m_isGBS, as it's
definitely false when handling wxStaticBoxSizer node.