Fix using slashes in wxFileDialog paths
Convert slashes to backslashes when using SHCreateItemFromParsingName()
as it doesn't support the former and returns E_INVALIDARG for the
strings containing them.

This fixes a regression in wxFileDialog::SetDirectory() since the switch
to using IFileDialog in wxMSW, as it didn't set the initial directory
correctly if it contained slashes, unlike before (because we already
replaced slashes with backslashes manually when using common file
dialogs).