Make generic wxListCtrl lines more visible
Use wxSYS_COLOUR_GRAYTEXT, as wxSYS_COLOUR_3DLIGHT is not visible at all
on the default white background under macOS and is barely visible with
the default GTK theme.

On the flip side, it's too bright in the dark themes, but this is,
arguably, better than not being visible at all.

See #22664.