Simplify the code using custom wxNSTextFieldEditor
Instead of triplicating (since the last commit, it was only duplicated
before) fieldEditor and setFieldEditor: methods in different classes and
then testing for these classes in wxNonOwnedWindowController code, use a
property with the special "WXFieldEditor" name to indicate that a class
should be using a wx editor and just test whether this property is
available in windowWillReturnFieldEditor:toObject:

This simplifies both the controls themselves and the code using the
editor and will allow adding it to other classes later if necessary.

No real changes, this should result in no changes in behaviour.