Fix images with alpha in wxMSW wxImageList without mask support
This was broken by the pre-3.2 changes to wxImageList which resulted in
not undoing alpha pre-multiplication when adding such images to the
image list.

In particular, this broke the appearance of disabled button images in
wxToolBar, as it uses wxImageList for storing them.

See #22754.

(cherry picked from commit efcb15f2bf33bbe6a7c9192695ab044801a16b6f)