Fix check for parameter validity in wxGrid::FreezeTo()
The number of rows or columns to freeze may be equal to the total number
of rows or columns, it just can't be strictly greater than it.

This is especially important when trying to thaw a grid which doesn't
have any rows, for example: before this change FreezeTo() would simply
return immediately, leaving the number of frozen columns unchanged and
quite possibly invalid (i.e. really out of range). Fix this by ensuring
that we reset it correctly now.

Also fix the types of the parameters in the documentation.