Don't log misleading errors if wxFileName::Rmdir() fails
In wxMSW, SHFileOperation() can return non-Win32 error codes when it
fails, but wxLogApiError() assumes that it's passed a Win32 error, which
resulted in nonsensical messages, such as "This function is not
supported on this system" for the error 0x78 which actually is
DE_ACCESSDENIEDSRC, i.e. "Security settings denied access to the
source".

It's better not to give any error message at all, in this debugging log
statement, than to give a wrong one, so log just the error code and let
people refer to the documentation to find out what it means.