Fix creating wxGLCanvas with NULL attributes list
Using NULL attributes list was not the same as using default
wxGLAttributes, but used a platform-dependent set of attributes that was
traditionally used by old wxWidgets versions for compatibility.

However these attributes didn't specify some important attributes,
notably depth buffer size, and relied on OpenGL driver using some
reasonable defaults, but this turns out to not be the case any more and
resulted in not enabling depth buffer at all, see #22787.

Fix this by using the same default attributes as were already used by
wxGLAttributes::Defaults() in this case, as this is much more useful,
even if not 100%-compatible with the old behaviour.

See #22828.