Avoid deprecation warnings for gdk_cairo_create()
We have to use this function for wxWindowDC implementation and setting
drag-and-drop icons, even though it is deprecated, as this API can't be
implemented otherwise, so suppress the deprecation warnings about doing
it.

Also include the private GTK header in the Cairo wxGraphicsContext
implementation just to avoid the deprecation warnings there too.