Fix return value of wxZoomGestureEvent::GetZoomFactor() in wxMSW
Although fingerDistance parameter is passed as a DWORD, it is actually
a signed integer which can be negative, so handle consistently handle it
as such to ensure that zoomFactor is computed correctly, as a ratio of 2
integers.

See #22818, #22819.

(cherry picked from commit 2e9b78c8329bcead9e2a03535d6fdff427b381f0)