Split RefreshArea() into two overload functions
One that refreshes part of one area of the grid and takes a wxRect
as parameter. the other one can refresh one or more areas entirely.

This makes sense and less confusing because a rectangle cannot be shared
with two separate/unrelated areas.

Notice that this commit undoes this 3bb86135e4 (Move wxGridArea enum
to private header)