wxGrid::Refresh() cleanup
This function expects the rectangle to be calculated relative to
the wxGrid window itself (with the origin (0, 0) in the top left
corner of the window) which is the caller's responsibility to make sure.

Additionally, wxScrolledCanvas::Refresh() will properly refresh the
intersection of subwindows with the rectangle rect. so no need to do
anything special inside the function which is at best an unnecessary
complication.