Better way to refresh the grid on wxGrid::Redimension()
Do not refresh the unaffected label window unnecessarily due to
the last call to Refresh().

In fact, we can do even better by refreshing the grid window the same way
wxGrid::DoSet{Row,Col}Size() do that. but for now let's continue with the
simple and straightforward way to refresh the grid.