Add serial number to wxwin.m4
This is a common convention and is helpful when installing this file,
see https://www.gnu.org/software/automake/manual/html_node/Serials.html

The serial value of 42 corresponds to the number of commits to this file
so far (and is also the right answer to the question of "what should the
serial be", of course, just as to any other question).

See #22848.

(cherry picked from commit aa06e0cbf97816f1f57947568978b837e41c3fb5)