Fix not getting some wxComboBox events after recreating it
Subclass the new EDIT control recreated together with the COMBOBOX
itself to make sure we still process the messages from it correctly.

This notably ensures that we still get wxEVT_TEXT_ENTER events in
wxBitmapComboBox under MSW after adding some items with images to it, as
this recreates the combobox internally.