Avoid warnings about deprecated functions inside glib.h
This happens when using gcc 4.8 under Ubuntu 18.04 and has to be
suppressed in the source file including the header as just doing it in
the header itself, with the usual wxGCC_WARNING_SUPPRESS/RESTORE macros,
has no effect, i.e. the warning is still given even if we bracket glib.h
inclusion with them.