Disable compositing when using wxDragImage
ImageList_BeginDrag() doesn't seem to be compatible with the use of
WS_EX_COMPOSITED, so turn it off when using it.

We probably could turn it back on again in EndDrag(), but for now keep
things simple and just fix the regression which resulted in wxDragImage
not working at all any more since d4ced54755 (Merge branch
'msw-double-buffer', 2022-10-15).