Add helper wxWithImages::GetImageBitmapFor()
This function is simpler to use than calling GetBitmapBundle() first and
then calling GetBitmapFor() on it, avoids creating a temporary
wxBitmapBundle unnecessarily when wxImageList is used and, most
importantly, is public and not protected, which will allow
wxGenericTreeCtrl to use it for the other image lists than the main one.

The slightly unwieldy name is due to trying to use "Image" in all public
functions of this class, which should hopefully minimize clashes with
any user-defined functions that may exist in the classes deriving from
it, without requiring the use of even uglier "WX" prefix.

No real changes yet.