Use wxGrid::Refresh() to correctly refresh grid subwindows
We shouldn't worry about how to refresh frozen windows anymore as
wxGrid::Refresh() properly and transparently refreshes them for us
if they are present. All we have to do is make sure the rectangle
is correctly expressed relative to the wxGrid window itself and not
the grid area it belongs to.