Fix using wxGCDC (Direct2D) with scrolled wxDC
Because since 4f9186f (Increase usable scrolling range in wxMSW by a
factor of 10,000, 2022-04-30) the shift of wxDC origin is not passed
through HDC so we need to apply it as an internal transform of
ID2D1RenderTarget. This transform is for internal purposes only and is
not exposed with GetTransform().

See #22480.
See #22914.