Remove another not really UTF-8 related FIXME comment
The question of which type should be used in the absence of the format
string still seems to be valid, but not really related to using (or not)
UTF-8.

Leave the comment, but remove the FIXME, as it almost certainly won't be
fixed (but hopefully all this code might get removed as part of #22880).

Also replace a leftover NULL with nullptr for consistency and to pacify
the code style check.