Don't use custom drawing for wxStaticBox by default any more
This is not necessary when double buffering is used, as there is no
flicker even without using our custom drawing code, so don't use it in
this case to save some GDI resources and minimize the probability of
running into bugs (like the one fixed in the previous commit) and other
side effects (which previously had to be disabled by explicitly setting
msw.staticbox.optimized-paint option to 0).

Unfortunately we still need to use this code just for allowing to change
the colour of the label for now.