Get rid of wxHAS_MSW_TASKDIALOG checks
All the necessary declarations should be available in the SDK headers of
any still supported compilers.

Note that we still need to keep the run-time checks because the task
dialog may be unavailable even in the latest OS versions when using old
comctl32.dll, as it happens if the application doesn't have a manifest
specifying the use of comctl32.dll v6.

See #22689.