Fix check for native command link button support
It's only available in comctl32.dll v6.10, not under Windows Vista or
later, i.e. it can be not available even in the latest Windows versions
if comctl32.dll v5.82 is used, as is the case for all applications not
specifying that they use comctl32.dll v6 in their manifest and, in
particular, in the applications not using a manifest at all.