Make wxQtSignalHandler a non-template class
To handle wx events the template parameter Handler should derive from
wxEvtHandler. This requirement is enforced now by making wxQtSignalHandler
a non-template class with an explicit wxEvtHandler member variable.

As a side effect, this will tremendously eliminate a bunch of unnecessary
template instantiation and reduce code bloat as a result.