Improve implementation of WX_ASSERT_FAILS_WITH_ASSERT_MESSAGE
Use SUCCEED() in it so passing tests are actually counted (before
running a test using this macro resulted in misleading output saying
that 0 asserts passed) and also use FAIL_CHECK(), instead of FAIL(), to
which CPPUNIT_FAIL() expanded before, to allow the rest of the test to
run even if this check fails, as missing assertions shouldn't be fatal.

Also remove a very old and outdated comment mentioning CppUnit.

Closes #22984.