Handle toggle() signal too in wxQtAction
Check the item in toggle() signal and send the event in triggered() signal
which emitted after.

We need this separation because triggered() is not emitted when setChecked()
or toggle() is called. The wxQtActionGroup (will be added in the next commit)
uses one of them to synchronize QActions with their wx wrappers and the menu
test will eventually pass.