Fix creating wxCairoContext from HDC (wxMSW)
HDC with negative x- or y-coordinate of the origin of the clipping region
is not supported by Cairo - either due to the Cairo bug
(prior to Cairo 1.15.12 - bug 96482) or by design (Cairo >= 1.15.12 -
https://gitlab.freedesktop.org/cairo/cairo/-/commit/4d07b57c168b88019a5510eaa7e9467149c53f12).
For such HDC we need to shift clipping region origin so that its coordinates
become >=0 and then pass this transformed HDC to Cairo surface.
To compensate this shift its value needs to be applied as an offset
added to the device coordinates when drawing to the surface.

See #22990.