Revert "Remove wxStaticBox custom drawing code entirely"
This reverts commit 616586f3011aa90c4bbced92d09e2c02a989d441 because it
turns out that using WS_EX_TRANSPARENT is incompatible with using
WS_EX_COMPOSITED (see parent commit).

Reintroducing all this code is painful, but there doesn't seem to be any
way of avoiding to have it as long as we use WS_EX_COMPOSITED. If we
stop using this style, we could revert this commit, this re-applying the
original change.

Closes #22982.