Simplify and correct setting URL in wxHyperlinkCtrl::Create()
It's unnecessary to call SetURL() twice and it's actually harmful to
call it before creating the window as it tries to update the label of
the window that doesn't exist yet (which results in a perfectly
avoidable warning when using Wine).

And it's unnecessary to call SetLabel() because we had just set the same
label anyhow. But we do need to take empty label into account when
either SetURL() or SetLabel() is called later, so move the check for
this from Create() into GetLabelForSysLink() helper, so that it is used
in all cases.

Finally, also remove the redundant SetVisited(false) as "visited" flag
is false by default.