Normalize displayed accelerators in wxMSW too
Show the standard accelerator representation instead of the original
string specified in the menu item label, as is already done in the other
ports.

This fixes the problem with showing "RawCtrl" in the menu for the
accelerators using it, as now "Ctrl" is displayed, as expected.

It also changes the display of the existing accelerators using "-", as
they're now displayed with "+" in wxMSW too, but this is already the
documented behaviour since the recent c4940c0838 (Update
wxMenuItem::SetItemLabel() documentation, 2022-11-24) and better
conforms to the platform standards -- and could also be (optionally?)
disabled later if it turns out to be really undesirable.

Closes #22892.

Closes #23017.