Deprecate unused wxTextDataObject::GetTextLength()
This function was supposed to be overridden in the subclasses to allow
efficiently implementing GetDataSize(), but it wasn't actually called
any longer since 229f00eb69 (Translate wxTextDataObject to/from native
EOL format., 2012-07-29).

And while it was probably a mistake to stop calling it in that commit,
the fact that nobody has complained about it since then seems to prove
that nobody was overriding it anyhow and it's not worth resurrecting it,
so just deprecate it instead.