Don't append extra NUL to URLs in wxGTK wxURLDataObject
The buffer size includes the trailing NUL, but we shouldn't make it part
of the string contents as it's not really part of the data at all.

The next commit will add a test verifying that retrieving an URL from
clipboard now gets the right string, without an extra trailing NUL, i.e.
fixes a bug similar to #22928 but in wxGTK.